Keep ManifestDefinition as part of Builder.
